requirements.txt
gitpython
pygithub
docker
requests